+++ src/adjusters/number.mjs
@@ -4,6 +4,7 @@ import Default from "../decorators/default";
 import AcceptNull from "../decorators/acceptNull";
 import AcceptEmptyString from "../decorators/acceptEmptyString";
 import Only from "../decorators/only";
+import Map from "../decorators/map";
 import Type from "../decorators/number/type";
 import MinValue from "../decorators/number/minValue";
 import MaxValue from "../decorators/number/maxValue";
@@ -20,6 +21,7 @@ export default () =>
 /**
  * adjuster for number
  */
+@Map
 @MaxValue
 @MinValue
 @Only