+++ src/index.d.ts
@@ -150,6 +150,13 @@ interface NumberAdjuster extends AdjusterBase<number>
 	 * @returns chainable instance
 	 */
 	maxValue(value: number, adjust?: boolean): this
+
+	/**
+	 * mapping
+	 * @param mapper mapping function
+	 * @returns chainable instance
+	 */
+	map(mapper: (value: number, fail: () => never) => number): this
 }
 
 interface StringAdjuster extends AdjusterBase<string>
@@ -215,6 +222,13 @@ interface StringAdjuster extends AdjusterBase<string>
 	 * @returns chainable instance
 	 */
 	pattern(pattern: RegExp): this
+
+	/**
+	 * mapping
+	 * @param mapper mapping function
+	 * @returns chainable instance
+	 */
+	map(mapper: (value: string, fail: () => never) => string): this
 }
 
 interface NumericStringAdjuster extends AdjusterBase<string>
@@ -274,6 +288,13 @@ interface NumericStringAdjuster extends AdjusterBase<string>
 	 * @returns chainable instance
 	 */
 	checksum(algorithm: string): this
+
+	/**
+	 * mapping
+	 * @param mapper mapping function
+	 * @returns chainable instance
+	 */
+	map(mapper: (value: string, fail: () => never) => string): this
 }
 
 interface EmailAdjuster extends AdjusterBase<string>